用户信息编辑模块：昵称修改接口，性别修改接口等实现。🍦

+++ src/main/java/cn/zjw/mrs/mapper/UserMapper.java
@@ -14,9 +14,9 @@ import java.util.List;
 */
 @Repository
 public interface UserMapper extends BaseMapper<User> {
-    List<String> selectUserTypes(Integer id);
+    List<String> selectUserTypes(Long id);
 
-    List<String> selectUserRegions(Integer id);
+    List<String> selectUserRegions(Long id);
 }
